+++ eth/main.cpp
@@ -37,6 +37,7 @@
 
 #include <libdevcore/FileSystem.h>
 #include <libdevcore/easylog.h>
+#include <libdevcore/StateMonitor.h>
 
 #include <libevm/VM.h>
 #include <libevm/VMFactory.h>
@@ -69,7 +70,7 @@
 #include <libweb3jsonrpc/Debug.h>
 #include <libweb3jsonrpc/Test.h>
 #include "Farm.h"
-
+#include <BuildInfo.h>
 //#include <ethminer/MinerAux.h>
 #include "AccountManager.h"
 
@@ -329,8 +330,8 @@ class ExitHandler: public SystemManager
 {
 public:
 	void exit() { exitHandler(0); }
-	static void exitHandler(int) { s_shouldExit = true; }
-	bool shouldExit() const { return s_shouldExit; }
+	static void exitHandler(int) { s_shouldExit = true; statemonitor::shouldExit = true; }
+	bool shouldExit() const { return s_shouldExit && !statemonitor::isRunning; }
 
 private:
 	static bool s_shouldExit;
@@ -511,7 +512,6 @@ int main(int argc, char** argv)
 
 	int jsonRPCURL = -1;
 	int jsonRPCSSLURL = -1;
-	bool adminViaHttp = true;
 	bool ipc = true;
 	std::string rpcCorsDomain = "";
 
@@ -734,7 +734,10 @@ int main(int argc, char** argv)
 		else if (arg == "-R" || arg == "--rebuild")
 			withExisting = WithExisting::Verify;
 		else if (arg == "-R" || arg == "--rescue")
+		{
 			withExisting = WithExisting::Rescue;
+			dev::g_withExisting = WithExisting::Rescue;
+		}
 		else if (arg == "--client-name" && i + 1 < argc)
 			clientName = argv[++i];
 		else if ((arg == "-a" || arg == "--address" || arg == "--author") && i + 1 < argc)
@@ -895,8 +898,6 @@ int main(int argc, char** argv)
 
 		else if ((arg == "-j" || arg == "--json-rpc"))
 			jsonRPCURL = jsonRPCURL == -1 ? SensibleHttpPort : jsonRPCURL;
-		else if (arg == "--admin-via-http")
-			adminViaHttp = true;
 		else if (arg == "--json-rpc-port" && i + 1 < argc)
 			jsonRPCURL = atoi(argv[++i]);
 		else if (arg == "--rpccorsdomain" && i + 1 < argc)
@@ -1387,9 +1388,6 @@ int main(int argc, char** argv)
 
 	auto caps = useWhisper ? set<string> {"eth", "shh"} : set<string> {"eth"};
 
-	//启动channelServer
-	ChannelRPCServer::Ptr channelServer = std::make_shared<ChannelRPCServer>();
-
 	//建立web3网络
 	dev::WebThreeDirect web3(
 	    WebThreeDirect::composeClientVersion("eth"),
@@ -1402,6 +1400,9 @@ int main(int argc, char** argv)
 	    testingMode
 	);
 
+	//启动channelServer
+	ChannelRPCServer::Ptr channelServer;
+	channelServer.reset(new ChannelRPCServer(), [](ChannelRPCServer *) {});
 	channelServer->setHost(web3.ethereum()->host());
 	web3.ethereum()->host().lock()->setWeb3Observer(channelServer->buildObserver());
 
@@ -1664,13 +1665,11 @@ int main(int argc, char** argv)
 	else
 		cout << "Networking disabled. To start, use netstart or pass --bootstrap or a remote host." << "\n";
 
-	unique_ptr<ModularServer<>> channelModularServer;
-	unique_ptr<ModularServer<>> jsonrpcHttpServer;
-	unique_ptr<ModularServer<>> jsonrpcHttpsServer;
-	unique_ptr<ModularServer<>> jsonrpcIpcServer;
 	unique_ptr<rpc::SessionManager> sessionManager;
 	unique_ptr<SimpleAccountHolder> accountHolder;
-
+	ModularServer<>* channelModularServer;
+	ModularServer<>* jsonrpcHttpServer;
+	ModularServer<>* jsonrpcIpcServer;
 	AddressHash allowedDestinations;
 
 	std::function<bool(TransactionSkeleton const&, bool)> authenticator;
@@ -1711,18 +1710,17 @@ int main(int argc, char** argv)
 
 		sessionManager.reset(new rpc::SessionManager());
 		accountHolder.reset(new SimpleAccountHolder([&]() { return web3.ethereum(); }, getAccountPassword, keyManager, authenticator));
-		auto ethFace = new rpc::Eth(*web3.ethereum(), *accountHolder.get());
-		rpc::TestFace* testEth = nullptr;
-		if (testingMode)
-			testEth = new rpc::Test(*web3.ethereum());
+		// rpc::TestFace* testEth = nullptr;
+		// if (testingMode)
+		// 	testEth = new rpc::Test(*web3.ethereum());
 
 		string limitConfigJSON = contentsString(chainParams.rateLimitConfig);
 
 		if (jsonRPCURL >= 0)
 		{
 			//no need to maintain admin and leveldb interfaces for rpc
-			jsonrpcHttpServer.reset(new FullServer(
-				ethFace,
+			jsonrpcHttpServer = new FullServer(
+				new rpc::Eth(*web3.ethereum(), *accountHolder.get()),
 				// new rpc::LevelDB(), new rpc::Whisper(web3, {}),
 				nullptr, nullptr,
 				new rpc::Net(web3),
@@ -1737,7 +1735,7 @@ int main(int argc, char** argv)
 				//adminEth, adminNet, adminUtils,
 				//new rpc::Debug(*web3.ethereum()),
 				//testEth
-			));
+			);
 			auto httpConnector = new SafeHttpServer("0.0.0.0", jsonRPCURL, "", "", SensibleHttpThreads);
 			httpConnector->setAllowedOrigin(rpcCorsDomain);
 			jsonrpcHttpServer->addConnector(httpConnector);
@@ -1754,8 +1752,8 @@ int main(int argc, char** argv)
 
 		if (ipc)
 		{
-			jsonrpcIpcServer.reset(new FullServer(
-			   ethFace,
+			jsonrpcIpcServer = new FullServer(
+			   new rpc::Eth(*web3.ethereum(), *accountHolder.get()),
 			   //new rpc::LevelDB(),
 			   //new rpc::Whisper(web3, {}),
 			   nullptr,
@@ -1773,7 +1771,7 @@ int main(int argc, char** argv)
 			   nullptr,
 			   nullptr,
 			   nullptr
-			));
+			);
 			auto ipcConnector = new IpcServer("geth");
 			jsonrpcIpcServer->addConnector(ipcConnector);
 			// jsonrpcIpcServer->setStatistics(new InterfaceStatistics(getDataDir() + "IPC", chainParams.statsInterval));
@@ -1782,9 +1780,9 @@ int main(int argc, char** argv)
 
 		//启动ChannelServer
 		if (!chainParams.listenIp.empty() && chainParams.channelPort > 0) {
-			channelModularServer.reset(
+			channelModularServer = 
 			    new FullServer(
-					ethFace,
+					new rpc::Eth(*web3.ethereum(), *accountHolder.get()),
 					nullptr, //new rpc::LevelDB(),
 					nullptr, //new rpc::Whisper(web3, { }),
 					new rpc::Net(web3),
@@ -1795,8 +1793,7 @@ int main(int argc, char** argv)
 					nullptr, //new rpc::AdminUtils(*sessionManager.get()),
 					nullptr, //new rpc::Debug(*web3.ethereum()),
 					nullptr //testEth
-				   )
-			);
+				   );
 
 			channelServer->setListenAddr(chainParams.listenIp);
 			channelServer->setListenPort(chainParams.channelPort);
@@ -1805,8 +1802,6 @@ int main(int argc, char** argv)
 			LOG(TRACE) << "channelServer started IP:" << chainParams.listenIp << " Port:" << chainParams.channelPort;
 
 			channelModularServer->StartListening();
-			//设置json rpc的accountholder
-            RPCallback::getInstance().setAccountHolder(accountHolder.get());
 		}
 
 		if (jsonAdmin.empty())
@@ -1873,15 +1868,12 @@ int main(int argc, char** argv)
 	}
 	
 
-	if (jsonrpcHttpServer.get())
+	if (jsonrpcHttpServer)
 		jsonrpcHttpServer->StopListening();
-	if (jsonrpcHttpsServer.get())
-		jsonrpcHttpsServer->StopListening();
-	if (jsonrpcIpcServer.get())
+	if (jsonrpcIpcServer)
 		jsonrpcIpcServer->StopListening();
-	if (channelModularServer.get())
+	if (channelModularServer)
 		channelModularServer->StopListening();
-
 	/*	auto netData = web3.saveNetwork();
 		if (!netData.empty())
 			writeFile(getDataDir() + "/network.rlp", netData);*/